Metrics Calculations on Uploaded Packages and Saving Metrics to DynamoDB
Implemented calculating metrics on uploaded content, or URLs for /package "POST". Runs metrics on both uploaded content and URLs, if the net score is above 0.5 then we continue with saving it, otherwise we return 424 Error "Package is not uploaded due to the disqualified rating.". If package is sucessfully uploaded meaning it passed the minimum net score then we also upload the metrics to our DynamoDB database where the partition key is the ID of the package/module.