Implemented user permissions and user groups
Added functionality to check for and handle user defined permissions (search, upload, download). Added functionality to upload a package with "secret" set to true or false which will then if true save your package to the database and attach an attribute call group which holds the group name that package belongs to.